+++ packages/orq-rc/src/models/operations/retrieveconversation.ts
@@ -193,7 +193,7 @@ export const RetrieveConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01ke9dbnv3wabypnsgws265h8w"),
+  _id: z.string().default("conv_01ke9hjvedx1mmc7v4rf73bm54"),
   entityId: z.string(),
   kind: RetrieveConversationKind$inboundSchema,
   displayName: z.string(),
@@ -226,7 +226,7 @@ export const RetrieveConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01ke9dbnv3wabypnsgws265h8w"),
+  id: z.string().default("conv_01ke9hjvedx1mmc7v4rf73bm54"),
   entityId: z.string(),
   kind: RetrieveConversationKind$outboundSchema,
   displayName: z.string(),