+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -307,7 +307,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01ke9dbnv55wq7hs42pz98z1n1"),
+  _id: z.string().default("conv_01ke9hjveeqkmy4hb9dvt879vc"),
   entityId: z.string(),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
@@ -341,7 +341,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01ke9dbnv55wq7hs42pz98z1n1"),
+  id: z.string().default("conv_01ke9hjveeqkmy4hb9dvt879vc"),
   entityId: z.string(),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),