+++ packages/orq-rc/RELEASES.md
@@ -5719,3 +5719,13 @@ Based on:
 - [typescript v4.2.0-rc.11] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.11 - packages/orq-rc
+
+## 2026-01-06 11:40:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.681.0 (2.789.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.12] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.12 - packages/orq-rc